gainloss/requirements.txt
gdax
pandas
requests
PyYAML